Experimental changes to the Birthing system
Changed class names from BirthsXXXX to XXXXBirths.
Births is now an abstract class.
Created Enum BirthNumber to replace the booleans firstBirth, secondBirth
and thirdBirth.
Wrote a test program, TestBirthNumberEnum to test the enum.
Removed some setters that are (maybe) unnecessary.

Variables for the number of nurses required for the different births
(and for the different birth numbers) have been removed. Instead, each
subclass of Births will implement a birth() method, and the tests for
the number of nurses required will be done there.

In the Resources class, changed Larves (French) to Larvae (English)