Different forms for some pheromones
Medium pheromones and Big pheromones now have two forms, either CLUSTER
or LINEAR.

Tiles have been repackaged in two subpackages: one for special tiles,
one for pheromones.

The classes Tile, SpecialTile and Pheromone are now abstract.

The subclasses implement a method validatePosition that will check to
see that the hexagons in the arguments correspond to the shape and size
of the tile.